Remove envelope when building the model from hash attributes
All the responses from the API have an envelope:

```
{
  "user": {
    "id": "64d707a8-a51e-4e3b-ad48-c4c33a73d176"
  }
}

```

The generated models expect the attributes at the top level like this:

```
{
  "id": "64d707a8-a51e-4e3b-ad48-c4c33a73d176"
}
```

Currently, grape-swagger does not understand the 'root' element when
generating the documentation. Because the generated documentation does
not match reality the generated models cannot corrently map the
responses to the attributes.

Apparently the 'root' attribute is not well supported and, in hindsight,
it was a bad choice. The envelope will be removed in a subsequent
version of the API.